Clear nurseries in parallel, with each capability's gc thread clearing its nursery.
clearNurseries() is only executes the sanity check code, since all the other code that was in clearNurseries() is performed in clearNursery(). To communicate the allocated amount in each cap's nursery back to the main GC thread, I added a field to struct gc_thread_ which is filled in by a worker thread after calling clearNursery(). The field is cleared in init_gc_thread().